fix: useSearchParam build error
- add a clientSideTransiton component and wrap with Suspense
